% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fba_tools.R
\name{get_exchange_reactions}
\alias{get_exchange_reactions}
\title{Get table of exchange reactions}
\usage{
get_exchange_reactions(mod, algorithm = "pfbaHeuristic")
}
\arguments{
\item{mod}{Model object of class `ModelOrg`}

\item{algorithm}{character. Algorithm to use to calculate flux distribution.}
}
\value{
data.table with the following columns: `ID` - identifiers of the
exchange reactions, `name` - name of the exchange reaction/compound, `flux` -
predicted flux (negative for uptake, positive for production), `lb` - lower
bound for flux, `ub` - upper bound for flux.
}
\description{
Get a table with all exchange reactions, their predicted fluxes,
and lower+upper bounds.
}
