% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{get_reaction_table}
\alias{get_reaction_table}
\title{Get table of reactions}
\usage{
get_reaction_table(mod)
}
\arguments{
\item{mod}{Model of class `ModelOrg`}
}
\value{
data.table with the columns: (`react.id`) Reactions IDs, (`react.name`)
Reaction names, (`equation`) the reactions' equations. If models were built
with gapseq the additional columns `EC`, `gs.origin`, `keggID`, and `metacycID`
are also part of the output table.
}
\description{
Produces a table with the model's reactions (ID, name, equation).
}
